Create our own exceptions roots
We have a set of exceptions that are in the PhpExtension, which are the
vendor exceptions. The exceptions specific to this vendor and which
are reused across several of this vendor projects. These are our vendor
root exceptions.

Then we also have the project specific exceptions, which extend the
vendor exceptions. This will allow us to distinguish, at runtime,
which are *our* exceptions and which are exceptions thrown by external
tools.

Furthermore, we use a trait so we can have multiple inheritance, and an
 empty interface matching the trait so that we can use `type-hinting`
 and `instanceof`.