Create our own top level Test classes
AbstractIntegrationTest

The integration tests are used to test the integration of the
application core with 3rd party tools and delivery mechanisms. They
need to boot the application and have the overhead of simulating http
requests so they are even slower than the Integration tests.

AbstractFunctionalTest

We created a top level abstract functional test class to hold
convenience methods.
A functional test will test the application as a black box, emulating
user http requests.
Usually, for this, we need to get services out of the container, etc.
This top class makes it easier.
Furthermore, as the integration tests, the functional tests need to
boot the application and have the overhead of simulating http requests
so they are even slower than the Integration tests.

Create a container to be used in the tests as a service locator

We need to get services out of the container, but since
SF4 they are private by default, so we need to create a compiler pass
to make them available in a special test container.