Segregate the blog/admin controller
A problem we tend to have with controllers is that they end up having
many action methods. Many times we don't even have a standard for what
to put in a controller. This ends up with a mess of controllers with
unrelated actions, which add to the unpredictability of the codebase.

To prevent this, we need to find a standard that will limit the amount
of actions in a controller and at the same time maintains those
actions cohesive and predictable.

Thus, we segregate the controllers in a RESTish fashion, where a
controller only acts on one resource and only has one action for each
HTTP method possible.

Of course, we can not fully implement REST in controllers that render
HTML, at least when we want to edit a resource, we need to show data
in a form and this is not a REST use case.
However, there are very few cases where we need to do this and it's
perfectly possible to create our own standard, for example having
controller action names that match the HTTP method names when
appropriate plus 'create', 'edit' and 'search' to show forms for
those purposes. Or an 'upsert' action that can be used both to get
the form that creates/edits a resource and to post the form data and
effectively make changes.