Containerize the application
Set up tools to easily use docker to run our application in prd, dev and
tst environments.

We use a base dockerfile to set up a base image for our project, with
all the system requirements as these are the ones changing less often.
This way we do not have to install all these dependencies every time
we build a new prd image to deploy.

prd is the production environment although running on the builtin php
web server, it is enough for the purposes of this project, for the
time being. We made "symfony/web-server-bundle" a requirement so that
we can run it for production.The generated container image is
completely self contained, it contains all, and only, the code needed
to run on prd. There is one volume configured so that we can have
persistent state between deployments (sqlite DB).

dev is used during development and includes xdebug. It also has a
volume configured so we can change the project files in the host and
have the container immediately run them without needing to create a
new container.

tst is the environment used to run the tests during development. As the
dev environment it uses a volume to run tests on the src files under
development and has xdebug configured so we can debug the tests.