Containerize the application
Set up tools to easily use docker to run our application in prd, dev,
tst and ci environments.

We use a base dockerfile to set up a base image for our project, with
all the system requirements as these are the ones changing less often.
This way we do not have to install all these dependencies every time
we build a new prd image to deploy.

prd is the production environment although running on the builtin php
web server, it is enough for the purposes of this project, for the
time being. We made "symfony/web-server-bundle" a requirement so that
we can run it for production.The generated container image is
completely self contained, it contains all, and only, the code needed
to run on prd. There is one volume configured so that we can have
persistent state between deployments (sqlite DB).

ci is the environment where we run the CI tests. It is based on the
prd image, although we need the tests folder and dependencies to
be able to run the tests. It does not have xdebug installed, as we can
use the php builtin debugger to generate the coverage reports.

dev is used during development and includes xdebug and an ssh server so
we can have the IDE connect to the container through ssh. It also has a
volume configured so we can change the project files used by the
container.

tst is the environment used to run the tests during development. As the
dev environment it uses a volume to run tests on the src files under
development and has xdebug configured so we can debug the tests.