Integrate with Scrutinizer CI
Now we introduce the `ci` environment, which is the environment
where we run the CI tests.

It is based on the prd image, although we need the tests folder
and dependencies to be able to run the tests.

It does not have xdebug installed, as we can use the php builtin
debugger to generate the coverage reports.