Create a Uuid and a UuidGenerator
The UuidGenerator is useful because it makes Uuid objects
predictable and therefore testable.

The Uuid value object can be used as the entities ID, as opposed to the
numeric ID generated by the DB engine. This means that the entities,
the core domain, can be decoupled/independent of the tool used for
persistence.

Without it, to create a complete entity, it always needs to be
persisted before getting an ID. So when we need the ID of an entity in
the same request as we create the entity, we will need to make a DB
call (flush) to persist the entity instead of making only one flush per
request, which is more lightweight for the system.

It will also help make testing easier, as we don't depend on a tool,
in order to have a fully functional entity.

Furthermore, when scaling the persistence mechanism to several servers,
if relying on the auto-increment IDs generated by the DB engine, we
will have problems because we will have several servers concurrently
generating unique and sequential IDs. However, if we don't need the
DB engine to generate the auto-increment IDs, we can safely scale the
amount of DB servers as much as we want.