Remove Comment entity from CommentCreatedEvent
Events should always be able to run async, ie using a message queue.
If we put complex objects in the event, it can bring problems when
serializing/deserializing them, or in the very least increase
complexity, when there is no need for such.

Entities are such complex objects, so instead of puting entities inside
events, it is preferred to put only the entities IDs.