Wrap the event listeners within a DB transaction
Since the events are going to be handled outside of the main use case
transaction, they will also be executed outside of the request
transaction (after it is closed) so we will need to open a new
transaction for the event listeners.
Furthermore, since each event is independent from each other, we need
to open a new transaction for each individual event listener.