Prevent race condition when generating sequent. ID
We don't rely on the DB engine to provide the sequential IDs anymore,
however we still need to have sequential IDs in our application(s).

To create the "next" ID we need to get the highest existent ID, figure
out the next one and save it.

The problem is that there might be several requests doing the same at
the same time, so they all get the same highest existent ID (ie 26) and
they all figure out the same "next one" (ie 27).
Then they all save it and we either have several IDs with the same
number, or the DB engine doesn't even allows saving those duplicate IDs.

The solution is to use a lock mechanism to prevent several requests
creating those sequential IDs at the same time. However, this also means
 that we will have requests waiting for other requests to finish, which
 is not a nice thing. So we need to be careful when using this approach
 and reduce to the maximum the risk of having this waiting queue.

In this specific case, we will only lock the post creation when the post
 being created has a slug which already exists and there is another
 request trying to create another one with the next sequential ID
 postfix. So the chances of it happening are low.

If we would lock all the new post creation requests however, that could
be a problem...

If the use case that has this locking mechanism is only triggered by a
console command however, for example when generating invoices, the
waiting time should not really be a problem because it is done async,
there is no user waiting for an immediate response.